modify world data so that it has a starting room
To make the player's starting room safe and never spawn enemies
there, I need to know where it is, but the way it was coded the player
spawn room was only decided after the world generation, and
there was no easy way to find its boundaries.

So I modified world_data and generate rooms to always have a "fixed"
starting room, the starting room size can be configured in exported
vars on generate_rooms node, and it's position will random.

I also modified world_data to be able to return "typed" arrays with cell
indexes of only the desired CellType. This will make it easier to spawn
enemies based on type of cells.