Populating crypts with sarcos on one "main wall" (#488)
* create sarcophagus scene that spawns lids on top of itself

And also some adjustments to sarcophagus base position and to
lids collision shape margins

* rename property on pillar requirement

* Improve SpawnData

Now you can set any arbitrary position, can add custom properties that
will be set to the spawned node once it is spawned, and spawn data
keeps track if it has already been spawned or not so that we can add
big objects that occupy more than on cell

* GenerateSarcophagus first draft

* Add doorways to room data

* adjust sarcophagus after rebase with dev

* first draft of sarco step with sarco width bigger than 1

* fix some errors in world_data and simplify "fill room"  function

* Create a debug room scene to test sarcophagus spawning

This commit actually creates a base "debug room" scene and then
an inherited scene from it to test the sarcophagus. This is a quick way
to create a single room the way we want and test anything needed on it.

* fix problems with sarcophagus spawn

* some fixes while writing PR