Refactor player spawn position
Because teh staircases takes all tiles from their rooms, we have to
spawn the player in specific good positions in relation to
each staircase.

So we need to change how we get that position, but since
spawning in general only happens after world generation, we
can only have that information after spawning all scenes, and
a lot ended up having to change.