Populate staircases (#519)
* Create staircase scenes

* Add staircase populating step and...

...remove ground tiles from down staircases on
generate_grid_tiles

* Refactor player spawn position

Because teh staircases takes all tiles from their rooms, we have to
spawn the player in specific good positions in relation to
each staircase.

So we need to change how we get that position, but since
spawning in general only happens after world generation, we
can only have that information after spawning all scenes, and
a lot ended up having to change.

* fix error I found in a specific seed when removing edges

it was possible to leave an empty array in delaunay dict

* remove initial light

It's not possible to find the player's spawning position before
spawning staircases, so this doesn't work here anymore.

This was just a light to help debugging, so I just removed it
as we can use the debug light in the filters, but if we want
to add it as a feature it might be better to have the player
start already holding one type of light

* remove a leftover debug print

* small change to room data

adds a public function that might be useful and makes the code
easier to read